+++ CHANGELOG.md
@@ -5,12 +5,17 @@
 ### Added
 
 - `Header` and `Panel` menu items now have have an boolean active state. ([KRAD-263])
+- Ability to add letters inside `Avatar`. ([KRAD-283])
 
 ### Changed
 
 - Change default font size of `Text` to `body.paragraph` variant. ([KRAD-246])
 - `LinkHeading` now contains it's `id`.
 - Theme font styles have been corrected and adjusted. ([KRAD-266])
+
+### Fixed
+
+- `HeaderAvatar` cell not pulling passing props. ([KRAD-283])
 
 ### Removed
 
@@ -20,6 +25,7 @@
 [krad-263]: https://akrade.atlassian.net/browse/KRAD-263
 [krad-246]: https://akrade.atlassian.net/browse/KRAD-246
 [krad-266]: https://akrade.atlassian.net/browse/KRAD-266
+[krad-283]: https://akrade.atlassian.net/browse/KRAD-283
 [krad-239]: https://akrade.atlassian.net/browse/KRAD-239
 [krad-245]: https://akrade.atlassian.net/browse/KRAD-245