+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Change default font size of `Text` to `body.paragraph` variant. ([KRAD-246])
 - `LinkHeading` now contains it's `id`.
 - Theme font styles have been corrected and adjusted. ([KRAD-266])
+- The push menu item animation has been reworked. ([KRAD-258])
 
 ### Fixed
 
@@ -25,6 +26,7 @@
 [krad-263]: https://akrade.atlassian.net/browse/KRAD-263
 [krad-246]: https://akrade.atlassian.net/browse/KRAD-246
 [krad-266]: https://akrade.atlassian.net/browse/KRAD-266
+[krad-258]: https://akrade.atlassian.net/browse/KRAD-258
 [krad-283]: https://akrade.atlassian.net/browse/KRAD-283
 [krad-239]: https://akrade.atlassian.net/browse/KRAD-239
 [krad-245]: https://akrade.atlassian.net/browse/KRAD-245