+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.11.0] - 2022-02-14
+
 ### Added
 
 - `Header` and `Panel` menu items now have have an boolean active state. ([KRAD-263])
@@ -372,7 +374,8 @@ Upgraded dependencies. ([KRAD-42])
 
 Initial development
 
-[unreleased]: https://github.com/akrade/krado-react/compare/v0.10.0...HEAD
+[unreleased]: https://github.com/akrade/krado-react/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/akrade/krado-react/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/akrade/krado-react/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/akrade/krado-react/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/akrade/krado-react/compare/v0.7.2...v0.8.0